Initial Implementation of ignore users
This commit creates three routes that allow a user to ignore, stop
ignoring, and view all the users a user ignores. A hook method was also
created to bind to posts.byThread.post that checks all the users to see
which one is ignored, and then adds a 'ignored' property to the user
property.

4 DB methods were also created to support the above.

A client folder is also included that adds a 'ignored users' view that
allows a user to view all the users that particular user is ignoring. It
also includes a directive that is used in the post view to show whether
a user is ignored, ignore a user, and jump to the ignored users view.